tasks done: +Fixed the singleitemview youtube player buffering problem when setting video to fullscreen or turning device to landscape (automatically sets video to fullscreen). Now does not buffer during these changes. +Merged the original singleitemview layout. +Added some dynamic url to test the viewpager screenslider
Answers to Comments:
- I've already configured this to allow multiple youtube videos to
play at once. (Though i dnt think that would be convenient for
the user.) I think we could just add a button to switch between videos
[English][Japanese]. And we will just add the corresponding option
in the usbong builder.
- Multiple images are stored in the viewpager. I'm sure we could handle
X amount of images dynamically. This will display the usbong icon while
it is still loading or is not found.
- Multiple videos on a viewpager is not possible i think. I added
a comment in the youtube fragment that discusses it.